Better pipeline layout to work with with_columns
This is close to the original, but we need to specify
the set of spark safe functions. This is okay-ish I think.
It just also requires that we include the pipeline in the
larger DAG too...
